Make required-literal scaling test deterministic
